backend/requirements.txt
# ======================
# --- Core Backend ---
# ======================
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.8.2
requests==2.32.3
httpx==0.27.0
# ...existing code...
pyannote.audio
# ...existing code...
# ======================
# --- Databases ---
# ======================
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
motor==3.4.0             # Async MongoDB driver
pymongo==4.6.1           # Must be pinned to match Motor


# ======================
# --- AI / NLP / Speech ---
# ======================
transformers==4.43.3
torch==2.3.1             # Will auto-detect GPU if available, fallback to CPU
torchaudio==2.3.1
numpy==1.26.4
scipy==1.13.1
scikit-learn==1.5.1
spacy==3.7.5
nltk==3.9.1
openai-whisper==20231117
pyannote.audio==3.3.1
openai==1.40.3

# ======================
# --- Dev Tools (Optional) ---
# ======================
black==24.4.2
isort==5.13.2